Private API

Index

GridSim.firstday_with_nbterr

GridSim.nb_territories

Documentation

GridSim.firstday_with_nbterr — Method

firstday_with_nbterr(grid_size::Tuple{Int,Int}, territory_size::Tuple{Int,Int}, fill::Vector{V})

Same as firstday but with the number of territories returned as well.

GridSim.nb_territories — Method

nb_territories(grid_size::Tuple{Int,Int}, terr_size::Tuple{Int,Int})

Return the number of territories in the grid for each axis.